Fix final annoyances that prevented compile under go 1.0.x
- Timeout handling is now based on "time" package
  + relative timeouts and intervals are now done by time.Duration
  + What previously has been a SetFooTimeout is now a SetFooDeadline
    and takes a fixed deadline instead of a relative timeout.

- os.EAGAIN and friends are now under syscall.EAGAIN and friends.
  Actually we will see other error classes and can check whether
  this is a temporary error instead. This still needs to be done!

- http.Request has no field RawURL, but a field named URL instead.

"go test" reports PASS now

+++ transport_jsonppolling.go
@@ -37,7 +37,7 @@ func jsonpPollingHijack(w http.ResponseWriter, req *http.Request, proceed func(S
 
 type jsonpPollingSocket struct {
 	net.Conn
-	i  uint
+	i  uint64
 	rb []byte
 }
